+++ tests/test_app.py
@@ -21,6 +21,12 @@ def test_explain():
 
         print(f"\nResponse:\n\n{response.payload['body']}")
 
+        assert response.payload['account'] is not None
+
+        account = json.loads(response.payload['account'])
+        assert account['email'] == 'unittest@polytest.ai'
+        assert account['org'] == 'polytest.ai'
+
 
 def test_explain_with_guideline():
     with Client(app) as client: